+++ src/visit/visit.ts
@@ -41,7 +41,7 @@ export function visitProgram(program: ts.Program, appOptions?: AppOptions): Pars
     for (const sourceFile of program.getSourceFiles()) {
         if (!sourceFile.isDeclarationFile && !sourceFile.fileName.endsWith('Type.ts')) {
             if (appOptions?.verbose) {
-                appOptions.log(`-> Visiting: ${path.relative(process.cwd(), sourceFile.fileName)}`)
+                appOptions.log(`👀 Visiting: ${path.relative(process.cwd(), sourceFile.fileName)}`)
             }
             ts.forEachChild(sourceFile, createVisit(checker, parsedLogics, sourceFile, appOptions))
         }
@@ -70,26 +70,43 @@ export function createVisit(
         const logicTypeName = `${logicName}Type`
 
         let logicTypeArguments = []
+        let logicTypeImported = false
 
+        // get "logicType" in "kea<logicType>(..)"
         const keaTypeArguments = ts.isCallExpression(node.parent) ? node.parent.typeArguments : []
         const keaTypeArgument = keaTypeArguments?.[0]
 
-        // // kea<logicType>(..)
         if (keaTypeArgument?.typeName?.escapedText === logicTypeName) {
             // kea<logicType<somethingElse>>(...)
             // store <somethingElse> on the generated type!
             if (keaTypeArgument.typeArguments && keaTypeArgument.typeArguments.length > 0) {
                 logicTypeArguments = (keaTypeArgument.typeArguments as ts.Node[]).map((a) => a.getFullText())
             }
+
+            // only if symbol resolves we mark the logic type as imported
+            const symbol = checker.getSymbolAtLocation(keaTypeArgument.typeName)
+            if (symbol) {
+                logicTypeImported = true
+            }
         }
 
         const pathString = getLogicPathString(appOptions, sourceFile.fileName)
+        let typeFileName = sourceFile.fileName.replace(/\.[tj]sx?$/, 'Type.ts')
+
+        if (appOptions?.rootPath && appOptions?.typesPath) {
+            const relativePathFromRoot = path.relative(appOptions.rootPath, typeFileName)
+            typeFileName = path.resolve(appOptions.typesPath, relativePathFromRoot)
+        }
 
         const parsedLogic: ParsedLogic = {
             checker,
+            node,
             logicName,
+            logicTypeImported,
+            logicTypeName,
             logicTypeArguments: logicTypeArguments,
             fileName: sourceFile.fileName,
+            typeFileName,
             actions: [],
             reducers: [],
             selectors: [],